Added README file for Concepts in SQL project
This commit includes the initial creation of the README.md file for my Concepts in SQL project. The README file outlines the project overview and learning objectives. It serves as documentation to explain the purpose of the project and the skills learned, including DML commands, transactions, views, triggers, and events in SQL